requirements.txt
biopython==1.76
matplotlib
matplotlib-venn
opencv-python
pandas
psutil
seaborn
scikit-learn
scipy